requirements.txt
torch~=2.7.1
torchvision~=0.22.1
matplotlib~=3.10.5
PyYAML~=6.0.2
numpy~=2.3.2
